#350 Improve GeD transactions.

+++ web/modules/custom/girchi_ged_transactions/src/Entity/GedTransaction.php
@@ -217,7 +217,7 @@ class GedTransaction extends ContentEntityBase implements GedTransactionInterfac
       ->setDescription(t('The time that the entity was last edited.'));
 
     // My custom fields.
-    $fields['ged_amount'] = BaseFieldDefinition::create('decimal')
+    $fields['ged_amount'] = BaseFieldDefinition::create('integer')
       ->setLabel(t('Ged amount'))
       ->setTranslatable(TRUE)
       ->setDisplayOptions('form', [
